% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BaseConstruction}
\alias{BaseConstruction}
\title{Construction of the base}
\usage{
BaseConstruction(db_post)
}
\arguments{
\item{db_post}{The database with the dummy variables. The antepenultimate 
and penultimate column must correspond to the response variable. 
The last column must correspond to the group (training or test) 
to which the individual has been assigned. It must be the output 
of the function "DataPreparation".}
}
\value{
The base obtained after the singular value decomposition 
of the density matrix. Only vectors whose singular values are 
greater than 0 are retained.
}
\description{
It gets the new base obtained after the singular value decomposition
}
