% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CalculateNewCoordinates}
\alias{CalculateNewCoordinates}
\title{Calculation of the new coordinates}
\usage{
CalculateNewCoordinates(db_post, base)
}
\arguments{
\item{db_post}{The database with the dummy variables. the antepenultimate 
and penultimate column corresponds to the response variable. 
The last column corresponds to the group (training or test) 
to which the individual has been assigned. It must be the output 
of the function "DataPreparation".}

\item{base}{The base obtained after the singular value decomposition 
of the density matrix}
}
\value{
The data frame with the coordinates in the new basis. The 
penultimate column corresponds to the response variable. 
The last column corresponds to the group (training or test).
}
\description{
It calculates the coordinates in the new basis of the vectors 
of the categorical co-variables
}
