% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CalculatePolarCoordinates}
\alias{CalculatePolarCoordinates}
\title{Calculation of the polar coordinates}
\usage{
CalculatePolarCoordinates(new_coordinates, db_post)
}
\arguments{
\item{new_coordinates}{The data frame with the coordinates in the new basis. 
The penultimate column must correspond to the response variable. 
The last column must correspond to the group (training or test). 
It must be the output of the function "CalculateNewCoordinates".}

\item{db_post}{The database with the dummy variables. The antepenultimate 
and penultimate column must correspond to the response variable. 
The last column must correspond to the group (training or test) 
to which the individual has been assigned. It must be the output 
of the function "DataPreparation".}
}
\value{
The data frame with the polar coordinates. The 
penultimate column corresponds to the response variable. 
The last column corresponds to the group (training or test).
}
\description{
It calculates the polar coordinates from the data frame containing 
the coordinate information in the new base.
}
