% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CategoricalDataClassificationAnalysis}
\alias{CategoricalDataClassificationAnalysis}
\title{Categorical Data Classification Analysis}
\usage{
CategoricalDataClassificationAnalysis(db, n_model, coordinates = "two")
}
\arguments{
\item{db}{Database predictor and response variables. 
It must contain either character data or factors. The last 
column must contain the response variable}

\item{n_model}{Number of individuals to build the map. The number 
of individuals in the test set is calculated by difference.}

\item{coordinates}{Coordinates used in the discrimination. 
By default, the two polar coordinates will be used ("two").
If you only want to do it for one, you must indicate the name of the 
coordinate ("theta" or "rho").}
}
\value{
CDCA_object. List of the results of the analysis: 
- preprocessed_db The database with the dummy variables. The antepenultimate 
and penultimate column corresponds to the response variable. 
The last column corresponds to the group (training or test) 
to which the individual has been assigned.
- base The base obtained after the singular value decomposition 
of the density matrix. Only vectors whose singular values are 
greater than 0 are retained.
- new_coordinates The data frame with the coordinates in the new basis. The 
penultimate column corresponds to the response variable. 
The last column corresponds to the group (training or test).
- polar_coor The data frame with the polar coordinates. The 
penultimate column corresponds to the response variable. 
The last column corresponds to the group (training or test).
- maxima Matrix with the maxima of the density functions of each 
class (rows) 
- prediction Data frame with the prediction made, the actual classification, 
and the score used for discrimination (the difference between the distance 
of the coordinates of each individual with each of the maxima of the 
two classes).
- results List of calculated predictive parameters (relative confussion
matrix, negative and positive predictive values, F1 score, Cohen's
kappa and Matthew's correlation coefficients).
}
\description{
It performs the Categorical Data Classification Analysis from a 
data.frame of categorical data
}
