% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ClassAssignment}
\alias{ClassAssignment}
\title{Assignment of the map-predicted class}
\usage{
ClassAssignment(coor, maxima_points, coordinates = "two")
}
\arguments{
\item{coor}{The data frame with coordinates of individuals. The penultimate 
column must correspond to the response variable. The last column must 
correspond to the group (training or test).It must be the output of 
the function "CalculateNewCoordinates" or "CalculatePolarCoordinates".}

\item{maxima_points}{A list from the function "MaximumDensityFunctions" 
with the maxima of the density functions of each class (class0 and
class1).}

\item{coordinates}{Coordinates used in the discrimination. 
By default, they are two coordinates, corresponding to the first 
two columns of the "coor" data frame. If you only want to do it for one, 
you must indicate the name of the column containing that coordinate.}
}
\value{
Data frame with the prediction made, the actual classification, 
and the score used for discrimination (the difference between the distance 
of the coordinates of each individual with each of the maxima of the 
two classes).
}
\description{
It calculates for each individual in the validation group ("test") the 
class predicted by the map. It builds a data frame with the prediction 
made, the actual classification, whether these are matched and the 
score used for discrimination.
}
