% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DataPreparation}
\alias{DataPreparation}
\title{Data preparation}
\usage{
DataPreparation(db, n_model)
}
\arguments{
\item{db}{Database predictor and response variables. 
The two last columns must contain the response variable.
The variables must be dummy encoded. You can use the makeNominalData function 
of the package ExPosition.}

\item{n_model}{Number of individuals to build the map. The number 
of individuals in the test set is calculated by difference.}
}
\value{
The database with the dummy variables. The antepenultimate 
and penultimate column corresponds to the response variable. 
The last column corresponds to the group (training or test) 
to which the individual has been assigned.
}
\description{
Transforms the data.frame with predictor and response 
variables into a data frame with dummy variables.
}
