% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_nps}
\alias{db_nps}
\title{db_nps}
\description{
Data frame
}
\details{
data frame containing information on 404 patients who had a solitary 
pulmonary nodule detected. It collects 8 predictor variables and a 
response variable (whether or not they were diagnosed with lung cancer 
after nodule detection).
}
\examples{
data(db_nps, package="CDCA")
}
\keyword{datasets}
