% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MaximumDensityFunctions}
\alias{MaximumDensityFunctions}
\title{Calculation of the maximum of the density functions of each class}
\usage{
MaximumDensityFunctions(coor, coordinates = "two")
}
\arguments{
\item{coor}{The data frame  with the coordinates for which you want to 
calculate the density maxima. The penultimate column must correspond to 
the response variable. The last column must correspond to the 
group (training or test). It must be the output of the function 
"CalculateNewCoordinates" or "CalculatePolarCoordinates".}

\item{coordinates}{Coordinates for which the maxima of density functions
are to be calculated. By default, it is calculated for two coordinates, 
corresponding to the first two columns of the "coor" data frame. 
If you only want to do it for one, you must indicate the name of 
the column containing that coordinate.}
}
\value{
Matrix with the maxima of the density functions of each class (rows)
}
\description{
It calculates the maximum of the density functions of the coordinate(s)
of the "model" group (training set) of each class. If two-coordinate maxima 
are calculated, the Gaussian kernel is used, if one-coordinate maxima 
are calculated, the Epanechnikov kernel is used.
}
